Seperating video loader code
