+++ app/src/main/java/io/smileyjoe/putio/tv/action/video/RefreshAction.java
@@ -5,7 +5,7 @@ import io.smileyjoe.putio.tv.network.Putio;
 import io.smileyjoe.putio.tv.object.Video;
 import io.smileyjoe.putio.tv.util.Async;
 import io.smileyjoe.putio.tv.util.PutioHelper;
-import io.smileyjoe.putio.tv.util.VideoLoader;
+import io.smileyjoe.putio.tv.video.VideoCache;
 
 public interface RefreshAction extends Action {
 
@@ -16,7 +16,7 @@ public interface RefreshAction extends Action {
             AppDatabase.getInstance(getContext()).videoDao().delete(getVideo().getPutId());
             PutioHelper helper = new PutioHelper(getContext());
             helper.setListener(video -> {
-                VideoLoader.update(getContext(), video);
+                VideoCache.getInstance().update(video);
                 update(video);
             });
             helper.parse(getVideo().getPutId(), getVideo().getParentTmdbId(), Putio.Files.get(getContext(), getVideo().getPutId()));
